+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 10.1.3 (Unreleased)
+## 10.1.3 (2023-06-23)
 
 - Add the ability to set roles and affiliations via the MUC occupant modal
 - Fix `isOnlyEmojis is not a function` when using only `@converse/headless`